% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizing_datasets.R
\name{basic_visualization}
\alias{basic_visualization}
\title{Teaching Basic Graph Visualization}
\usage{
basic_visualization()
}
\description{
This function teaches the user how to do a basic graph with \code{ggplot2}. It teaches them that they need to input data, and input what will be the x and y axis
for a scatterplot.
}
\examples{
\dontrun{
#basic_visualization will ask for user input in the console, and loop without that input.
basic_visualization()
}
}
