% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_correct.R
\name{check_if_correct}
\alias{check_if_correct}
\title{Check to see if the user input is correct.}
\usage{
check_if_correct(typed, correct)
}
\arguments{
\item{typed}{input from user via console}

\item{correct}{what we want the user to have typed}
}
\description{
`check_if_correct` compares the parameter `typed` to the parameter `correct`.
}
\details{
If the parameter `typed` is identical to the parameter `correct`, the function ends and the user can continue.
This causes the previous function (which called `check_if_correct`) to resume.
However, if the parameters are different, the function asks the user to reenter the value.
This calls the function again, to compare the new value, now the new `typed`, to the same `correct`.
The function continues until `typed` is identical to `correct`.
}
