% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizing_datasets.R
\name{downloading_ggplot2}
\alias{downloading_ggplot2}
\title{Downloading ggplot2}
\usage{
downloading_ggplot2()
}
\description{
Teaches user how to download the package \code{ggplot2}.
}
\examples{
downloading_ggplot2()
}
