% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_with_datasets.R
\name{filter_row}
\alias{filter_row}
\title{Filter Row}
\usage{
filter_row()
}
\description{
The function \code{filter_row} shows the user how to select filter rows using the filter() function.
}
\examples{
\dontrun{
#filter_row() will ask for user input in the console, and loop without that input.
filter_row()
}
}
