% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingbasics.R
\name{load_dplyr}
\alias{load_dplyr}
\title{Load dplyr}
\usage{
load_dplyr()
}
\description{
The function \code{load_dplyr} guides the user through the installation of \code{dplyr} (if neccessary) and loading the \code{dplyr} package.
}
\details{
\code{load_dplyr} asks the user if they've ever loaded the \code{dplyr} package before, and if not, guides them through the installation.
After \code{dplyr} has been installed, or if it has been loaded before,
the user is taught how to load the package using \code{library()}.
}
\examples{
load_dplyr()

}
