% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingbasics.R
\name{load_tidyverse}
\alias{load_tidyverse}
\title{Load tidyverse}
\usage{
load_tidyverse()
}
\description{
The function \code{load_tidyverse} guides the user through the installation of \code{tidyverse} (if neccessary) and loading the \code{tidyverse} package.
}
\details{
After \code{tidyverse} has been installed, or if it has been loaded before,
the user is taught how to load the package using \code{library()}.
}
