% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingbasics.R
\name{see_data_console}
\alias{see_data_console}
\title{See Data In Console}
\usage{
see_data_console()
}
\value{
country_indicators
}
\description{
\code{see_data_console} teaches the user how to view their data in the console.
}
\examples{
<<<<<<< HEAD
\dontrun{
#see_data_console will run in the console, where user input is needed.
see_data_console()
}

=======
see_data_console()
>>>>>>> daa27e4ed1021f95abd38b4a4218b00f853ca766

}
