% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_with_datasets.R
\name{select_column}
\alias{select_column}
\title{Select Column}
\usage{
select_column()
}
\description{
The function \code{select_column} shows the user how to select particular columns using the select() function.
}
