% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{start_package}
\alias{start_package}
\title{Beginning Interface}
\usage{
start_package()
}
\description{
\code{start_package} begins the user interface and introduces the package.
}
\details{
Ideally, the user should call \code{start_package} immediately upon loading the package,
in order to begin using socialsciencer.
While other functions can be called individually,
this is the one which should be called first.
There are no inputs or outputs.
}
\examples{
\dontrun{
#start_package() will ask for user input in the console, and loop without that input.
start_package()
}
}
