% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_correct.R
\name{check_if_correct}
\alias{check_if_correct}
\title{Check to see if the user input is correct}
\usage{
check_if_correct(typed, correct)
}
\arguments{
\item{typed}{input from user via console}

\item{correct}{what we want the user to have typed}
}
\description{
Check to see if the user input is correct
}
