% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling_functions.R
\name{check_if_correct}
\alias{check_if_correct}
\title{Check to see if the user input is correct}
\usage{
check_if_correct(typed, correct)
}
\arguments{
\item{correct}{}
}
\description{
Check to see if the user input is correct
}
