% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_correct.R
\name{check_if_correct}
\alias{check_if_correct}
\title{Check to see if the user input is correct.}
\usage{
check_if_correct(typed, correct)
}
\arguments{
\item{typed}{input from user via console}

\item{correct}{what we want the user to have typed}
}
\description{
\code{check_if_correct()} compares the parameter \code{typed} to the parameter \code{correct}.
}
\details{
If the parameter \code{typed} is identical to the parameter \code{correct}, the function ends and the user can continue.
This causes the previous function (which called \code{check_if_correct}) to resume.
However, if the parameters are different, the function asks the user to reenter the value.
This calls the function again, to compare the new value, now the new \code{typed}, to the same \code{correct}.
The function continues until \code{typed} is identical to \code{correct}.
}
\examples{
check_if_correct("x", "x")
\dontrun{
#Since "x" and "y" aren't equal, users should expect this to call check_if_correct() again.
check_if_correct("x", "y")
}


}
