% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizing_datasets.R
\name{downloading_ggplot2}
\alias{downloading_ggplot2}
\title{Downloading ggplot2}
\usage{
downloading_ggplot2()
}
\description{
Teaches user how to download the package \code{ggplot2}.
}
\examples{
\dontrun{
#downloading_ggplot2 will ask for user input in the console, and loop without that input.
downloading_ggplot2()
}
}
