% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_with_datasets.R
\name{filter_row}
\alias{filter_row}
\title{Title}
\usage{
filter_row()
}
\description{
Title
}
\examples{
\dontrun{
#filter_row() will ask for user input in the console, and loop without that input.
filter_row()
}
}
