% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dplyr.R
\name{load_dplyr}
\alias{load_dplyr}
\title{Load dplyr}
\usage{
load_dplyr()
}
\description{
The function `load_dplyr` guides the user through the installation of `dplyr` (if neccessary) and loading the `dplyr` package.
}
\details{
`load_dplyr` asks the user if they've ever loaded the `dplyr` package before, and if not, guides them through the installation.
After `dplyr` has been installed, or if it has been loaded before,
the user is taught how to load the package using `library()`.
}
