% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingbasics.R
\name{load_dplyr}
\alias{load_dplyr}
\title{Load dplyr}
\usage{
load_dplyr()
}
\description{
The function \code{load_dplyr} guides the user through the installation of \code{dplyr}  and loading the \code{dplyr} package.
}
\details{
\code{load_dplyr} guides users through the installation of dplyr using the \code{install.packages()} function.
After \code{dplyr} has been installed,
the user is taught how to load the package using \code{library()}.
}
\examples{
\dontrun{
#load_dplyr() will run in the console, where user input is needed.
load_dplyr()
}

}
