% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingbasics.R
\name{view_data}
\alias{view_data}
\title{View() Data}
\usage{
view_data()
}
\value{
utils::View(SocialScienceR::country_indicators)
}
\description{
\code{view_data} teaches the user how to see data using the View() function.
}
\examples{
\dontrun{
#view_data will ask for user input in the console, and loop without that input.
view_data()
}

}
